Add preliminary support for decomposing large CAS objects
This change adds support for a custom hashing algorithm called
BLAKE3ZCC. BLAKE3ZCC is a slightly modified version of BLAKE3:

https://github.com/BLAKE3-team/BLAKE3/

BLAKE3ZCC stands for BLAKE3 Zero Chunk Counter; it is identical to
BLAKE3, except that the Chunk Counter (CC) is always set to zero while
hashing data. This makes it possible to perform in-file deduplication
and to decompose large objects into smaller ones. Smaller objects may be
located by loading a manifest that is also stored in the CAS.

This patchset is incomplete and hasn't received a lot of testing. Making
it more complete and ready to be merged to master is left as an exercise
to the reader...